Fixed issue where sequential tasks failed to run
The code will now wait until the previous PID task has been run before executing the next. This is an attempt at using command based programming with a single file. The PID task is ran and then stopped as it reaches it set point. This then executes the next instance of the method in line in the autonomous task, giving a new setpoint, and the PID code runs again.